fix(tree): fix tree viewport to snap to current move when current move changes
